src/third_party/blink/web_tests/external/wpt/service-workers/service-worker/fetch-canvas-tainting-video-with-range-request.https-expected.txt

This is a testharness.js-based test.
PASS range responses from single origin (same-origin)
FAIL range responses from multiple origins (cross-origin first) assert_equals: expected "TAINTED" but got "NOT_TAINTED"
FAIL range responses from single origin with both opaque and non-opaque responses assert_equals: expected "TAINTED" but got "NOT_TAINTED"
Harness: the test ran to completion.
